# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ewura Hub API - Airtime & Data selling platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  agent: "agent",
  admin: "admin",
} as const;

export interface User {
  id: string;
  username: string;
  email: string;
  phone: string;
  role: UserRole;
  isVerified: boolean;
  walletBalance: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  user: "user",
  agent: "agent",
} as const;

export interface RegisterBody {
  /** @minLength 3 */
  username: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  emailOrPhone: string;
  password: string;
}

export type ProductNetwork =
  (typeof ProductNetwork)[keyof typeof ProductNetwork];

export const ProductNetwork = {
  MTN: "MTN",
  Telecel: "Telecel",
  AirtelTigo: "AirtelTigo",
} as const;

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

export const ProductType = {
  airtime: "airtime",
  data: "data",
} as const;

export interface Product {
  id: string;
  network: ProductNetwork;
  type: ProductType;
  name: string;
  description?: string;
  userPrice: number;
  agentPrice: number;
  value: string;
}

export type OrderNetwork = (typeof OrderNetwork)[keyof typeof OrderNetwork];

export const OrderNetwork = {
  MTN: "MTN",
  Telecel: "Telecel",
  AirtelTigo: "AirtelTigo",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  airtime: "airtime",
  data: "data",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  paystack: "paystack",
  wallet: "wallet",
} as const;

export interface Order {
  id: string;
  userId: string;
  username: string;
  network: OrderNetwork;
  type: OrderType;
  productName: string;
  recipientPhone: string;
  amount: number;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  paymentReference?: string;
  createdAt: string;
}

export interface OrderWithPayment {
  order: Order;
  paymentUrl?: string;
  reference?: string;
  message: string;
}

export interface OrdersResponse {
  orders: Order[];
  total: number;
  page: number;
  pages: number;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  paystack: "paystack",
  wallet: "wallet",
} as const;

export interface CreateOrderBody {
  productId: string;
  recipientPhone: string;
  paymentMethod: CreateOrderBodyPaymentMethod;
}

export interface Wallet {
  balance: number;
  userId: string;
  totalFunded: number;
  totalSpent: number;
}

export interface FundWalletBody {
  /** @minimum 1 */
  amount: number;
}

export type WalletTransactionType =
  (typeof WalletTransactionType)[keyof typeof WalletTransactionType];

export const WalletTransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface WalletTransaction {
  id: string;
  type: WalletTransactionType;
  amount: number;
  description: string;
  reference: string;
  createdAt: string;
}

export interface TransactionsResponse {
  transactions: WalletTransaction[];
  total: number;
  page: number;
  pages: number;
}

export interface InitPaymentBody {
  orderId: string;
}

export interface PaystackInitResponse {
  authorizationUrl: string;
  accessCode: string;
  reference: string;
}

export interface PaymentVerifyResponse {
  status: string;
  message: string;
  order?: Order;
}

export interface AdminStats {
  totalOrders: number;
  completedOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  totalAgents: number;
  verifiedAgents: number;
  totalUsers: number;
  recentOrders: Order[];
}

export interface AgentsResponse {
  agents: User[];
  total: number;
  page: number;
  pages: number;
}

export interface VerifyAgentBody {
  isVerified: boolean;
}

export type AdjustBalanceBodyType =
  (typeof AdjustBalanceBodyType)[keyof typeof AdjustBalanceBodyType];

export const AdjustBalanceBodyType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface AdjustBalanceBody {
  amount: number;
  type: AdjustBalanceBodyType;
  reason: string;
}

export type GetProductsParams = {
  network?: GetProductsNetwork;
  type?: GetProductsType;
};

export type GetProductsNetwork =
  (typeof GetProductsNetwork)[keyof typeof GetProductsNetwork];

export const GetProductsNetwork = {
  MTN: "MTN",
  Telecel: "Telecel",
  AirtelTigo: "AirtelTigo",
} as const;

export type GetProductsType =
  (typeof GetProductsType)[keyof typeof GetProductsType];

export const GetProductsType = {
  airtime: "airtime",
  data: "data",
} as const;

export type GetOrdersParams = {
  status?: GetOrdersStatus;
  network?: string;
  limit?: number;
  page?: number;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export type GetWalletTransactionsParams = {
  limit?: number;
  page?: number;
};

export type GetAdminOrdersParams = {
  status?: string;
  network?: string;
  limit?: number;
  page?: number;
};

export type GetAdminAgentsParams = {
  verified?: boolean;
  limit?: number;
  page?: number;
};