# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ewura Hub API - Airtime & Data selling platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/fund:
    post:
      operationId: fundWallet
      tags: [wallet]
      summary: Fund agent wallet (initiates Paystack payment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundWalletBody"
      responses:
        "200":
          description: Payment initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaystackInitResponse"
  /wallet/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      summary: Get wallet transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsResponse"
  /payments/initialize:
    post:
      operationId: initializePayment
      tags: [payments]
      summary: Initialize a Paystack payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitPaymentBody"
      responses:
        "200":
          description: Payment initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaystackInitResponse"
  /payments/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a Paystack payment
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        network:
          type: string
          enum: [MTN, Telecel, AirtelTigo]
        type:
          type: string
          enum: [airtime, data]
        productName:
          type: string
        recipientPhone:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, processing, completed, failed]
        paymentMethod:
          type: string
          enum: [paystack, wallet]
        paymentReference:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, username, network, type, productName, recipientPhone, amount, status, paymentMethod, createdAt]
...
    FundWalletBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
      required: [amount]
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        description:
          type: string
        reference:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, amount, description, reference, createdAt]
    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
      required: [transactions, total, page, pages]
    InitPaymentBody:
      type: object
      properties:
        orderId:
          type: string
      required: [orderId]
    PaystackInitResponse:
      type: object
      properties:
        authorizationUrl:
          type: string
        accessCode:
          type: string
        reference:
          type: string
      required: [authorizationUrl, accessCode, reference]
    PaymentVerifyResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        order:
          $ref: "#/components/schemas/Order"
      required: [status, message]
...