# lib/api-zod/src/generated/types/fundWalletBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ewura Hub API - Airtime & Data selling platform
 * OpenAPI spec version: 0.1.0
 */

export interface FundWalletBody {
  /** @minimum 1 */
  amount: number;
}